Updates after testing

+++ scripts/extract.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -ex
+#!/bin/sh -e
 
 ############################################################################
 #
@@ -14,28 +14,32 @@
 SCRIPT_DIR="$( cd "$( dirname "$0" )" && pwd )"
 . "${SCRIPT_DIR}"/functions.sh
 
-GIT_REPO_URL="$(getGitRepoUrl)"
-setupGit
+if [ -z "$NO_GIT" ]; then
+  GIT_REPO_URL="$(getGitRepoUrl)"
+  setupGit
 
-if [ -z "${CICD_BRANCH}" ]; then
-  CICD_BRANCH=main
+  if [ -z "${CICD_BRANCH}" ]; then
+    CICD_BRANCH=main
+  fi
+  echo "Checkout branch: ${CICD_BRANCH}"
+  git checkout -B "${CICD_BRANCH}"
+
+  # Delete BPM artifacts to propagate deletions from Web Modeller
+  git rm --ignore-unmatch "${MODEL_PATH}"/*.bpmn
+  git rm --ignore-unmatch "${MODEL_PATH}"/*.dmn
+  git rm --ignore-unmatch "${MODEL_PATH}"/*.form
 fi
-echo "Checkout branch: ${CICD_BRANCH}"
-git checkout -B "${CICD_BRANCH}"
 
-# Delete BPM artifacts to propagate deletions from Web Modeller
-git rm --ignore-unmatch "${MODEL_PATH}"/*.bpmn
-git rm --ignore-unmatch "${MODEL_PATH}"/*.dmn
-git rm --ignore-unmatch "${MODEL_PATH}"/*.form
+python "${SCRIPT_DIR}"/extract.py "$@"
 
-python "${SCRIPT_DIR}"/extract.py
+if [ -z "$NO_GIT" ]; then
+  git add -- *.bpmn  2>/dev/null || true
+  git add -- *.dmn  2>/dev/null || true
+  git add -- *.form  2>/dev/null || true
+  git add -- config.*  2>/dev/null || true
+  git status
 
-git add -- *.bpmn  2>/dev/null || true
-git add -- *.dmn  2>/dev/null || true
-git add -- *.form  2>/dev/null || true
-git add -- config.*  2>/dev/null || true
-git status
+  git commit -m "$(getCommitMessage)"
 
-git commit -m "$(getCommitMessage)"
-
-git push "${GIT_REPO_URL}" "${CICD_BRANCH}"
+  git push "${GIT_REPO_URL}" "${CICD_BRANCH}"
+fi
